Add samples and docs
