+++ sdk/webpubsub/Microsoft.Azure.WebPubSub.AspNetCore/tests/Samples/WebPubSubSample.HelloWorld.cs
@@ -1,17 +1,20 @@
-﻿using Microsoft.AspNetCore.Builder;
-using System;
-using System.Collections.Generic;
-using System.Text;
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using Microsoft.AspNetCore.Builder;
+using System.Threading.Tasks;
 
 namespace Microsoft.Azure.WebPubSub.AspNetCore.Tests.Samples
 {
-    class WebPubSubSample
+    public class WebPubSubSample
     {
         public void Configure(IApplicationBuilder app)
         {
+            #region Snippet:WebPubSubValidationOptions
             var wpsHandler = new WebPubSubRequestBuilder()
                 .AddValidationOptions(new WebPubSubValidationOptions("<connection-string>"))
                 .Build();
+            #endregion
 
             app.UseEndpoints(endpoints =>
             {
@@ -21,6 +24,26 @@ namespace Microsoft.Azure.WebPubSub.AspNetCore.Tests.Samples
                     await wpsHandler.HandleRequest(context, testHub);
                 });
             });
+        }
+
+        private sealed class SampleHub : ServiceHub
+        {
+            #region Snippet:WebPubSubConnectMethods
+            public override Task<ServiceResponse> Connect(ConnectEventRequest request)
+            {
+                var response = new ConnectResponse
+                {
+                    UserId = request.ConnectionContext.UserId
+                };
+                return Task.FromResult<ServiceResponse>(response);
+            }
+            #endregion
+
+            public override Task<ServiceResponse> Message(MessageEventRequest request)
+            {
+                var response = new MessageResponse("ack");
+                return Task.FromResult<ServiceResponse>(response);
+            }
         }
     }
 }
